Refactor SSO redirect middleware and simplify plugin configuration
