+++ tutor-sso-redirect/MANIFEST.in
@@ -1,4 +1,3 @@
 include README.md
 include LICENSE
 recursive-include tutorssoredirect/patches *.yml
-recursive-include tutorssoredirect/openedx *
