--- tutor-sso-redirect/tutorssoredirect/openedx/tutorssoredirect/__init__.py
@@ -1,1 +1,0 @@
-# SSO Redirect Middleware Package
